% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ersphere.R
\name{ersphere}
\alias{ersphere}
\title{Generates equally distributed random points on a sphere}
\usage{
ersphere(n, dim, acc, signs = NULL, initial.points = NULL)
}
\arguments{
\item{n}{The number of points to be generated.}

\item{dim}{The number of dimensions/variables}

\item{acc}{The precision (the number of proposed generated points).}

\item{signs}{The signs of the generated variables; set a vector with 1s or -1s where each element corresponds to one variable/dimension; e.g., if the points in two dimensional space are going to be generated, then the vector with two elements can be provided. If the first element would be 1 and the second element would be -1, then all values in the fist column will be positive, while all the values in the second column will be negative; if you do not care about the signs of all variables, then set \code{signs=NULL} (privzeto); if you do not care about the sign of only one variable, then set the corresponding element \code{signs} to \code{NA}.}

\item{initial.points}{Provide a matrix with the previusly generated points if you only want to add/generated some new/additional ones.}
}
\value{
\itemize{
\item \code{points} - Generated values (data frame with n rows and dim columns).
\item \code{posterior} - Posterior probabilities (the values of the Fisher's calsification linear discrimination function).
\item \code{max.min.dist} - Maximum minimal distance between the points.
}
}
\description{
It is used to generate random points on a circumference (two dimensions) or on a sphere (three dimentions) ... The points are more equally distributed as in the case of other approaches.
}
\seealso{
\code{rsphere}
}
\author{
Marjan Cugmas
}
