% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.network.LE.R
\name{gen.network.LE}
\alias{gen.network.LE}
\title{Relocating Links algorithm (RL algorithm)}
\usage{
gen.network.LE(
  BM = BM,
  LE = 0.4,
  n = 24,
  size = rep(1/nrow(BM), nrow(BM)),
  symmetric = FALSE
)
}
\arguments{
\item{BM}{An image matrix of a blockmodel.}

\item{LE}{Desired level of errors.}

\item{n}{Network size.}

\item{size}{A vector with the values specifying the clusters' sizes (the elements must sum to 1).}

\item{symmetric}{Wheter a symmetric network should be generated.}

\item{k}{Number of iterations.}
}
\value{
A network with selected blockmodel type and level of errors.
}
\description{
Generate network with a selected blockmodel and the level or errors.
}
\examples{
# generate initial and ideal network
cohesiveBM <- rbind(c("com", "nul"), c("nul", "com"))
network <- gen.network.LE(BM = cohesiveBM, LE = 0.5, n = 12, size = rep(0.5, 2))
}
\author{
Marjan Cugmas
}
