% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.RF.kmeans.R
\name{get.RF.kmeans}
\alias{get.RF.kmeans}
\title{Calculate the value of the Relative Fit measure for k-means blockmodeling approach}
\usage{
get.RF.kmeans(network, nSim, groups, implFun, nCores = 0, ...)
}
\arguments{
\item{network}{The (binary) network in a matrix form.}

\item{nSim}{The number of randomized networks.}

\item{implFun}{The name of the function to be used for blockmodeling.
Currently implemented are options \code{"omkm"} (from package \code{dBlockmodeling}) and
\code{"kmBlockORPC"} (from package \code{kmBlock}).}

\item{nCores}{The number of cores to be used by funcion \code{kmBlockORPC}.}

\item{\dots}{Additional parameters for function \code{omkm} or \code{kmBlockORPC}.}
}
\description{
The function calculate the value of the RF measure.
}
\author{
Marjan Cugmas
}
