% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.RF.R
\name{get.RF}
\alias{get.RF}
\title{Calculate the value of the Relative Fit measure for generalized blockmodeling approach}
\usage{
get.RF(
  network,
  nSim = 10,
  nCores = 0,
  groups = groups,
  bmType = list(cohes = cohes, ccp = ccp),
  symmetrize = FALSE,
  blockTypeWeights = 1,
  rep = 200
)
}
\arguments{
\item{network}{The (binary) network in a matrix form.}

\item{nSim}{The number of randomized networks.}

\item{nCores}{The number of cores to be used by funcion \code{kmBlockORPC}.}

\item{groups}{The number of clusters in a blockmodel (single value of a vector of values).}

\item{bmType}{The list of image matrices.}
}
\description{
It calculates the value of the Relative Fit measure.
}
\note{
The networks are randomized in such a way that loops are not allowed, but the density is
still the same as in the empirical networks (the degrees are not preserved)!
For asymmetric networks only!
}
\seealso{
\code{get.RF.kmeans}
}
\author{
Marjan Cugmas
}
