% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IB.R
\name{IB}
\alias{IB}
\title{The number of inconsistent blocks}
\usage{
IB(IM1, IM2)
}
\arguments{
\item{MI1}{The first image matrix. The possible block types are \code{com}, \code{nul} and \code{dnc}.}

\item{MI2}{The second image matrix. The possible block types are \code{com}, \code{nul} and \code{dnc}.}
}
\value{
A list with \code{IB} which is a number of inconsistent blocks and \code{order} which is an order of
groups from the original image matrix, corresponding to the number of inconsistent blocks.
}
\description{
The function calculates the number of incosnstent blocks between two image matrices.
}
\author{
Marjan Cugmas
}
