\name{internal}
\alias{internal}
\alias{moduleNetwork.aux}
\alias{connectModules}
\alias{fitBUM}
\alias{dbum}
\alias{rbum}
\alias{qbum}
\alias{pbum}
\alias{getComponent}
\alias{inv.logit}
\alias{logit}
\alias{bum.negLogLik}
\alias{bum.mle}
\alias{bum.EM}
\alias{qqbum}
\alias{bum.histogram}
\alias{bum.dalt}
\alias{bum.palt}
\alias{bum.qalt}
\alias{bum.ralt}
\alias{nem.featureselection}
\alias{createBN}
\alias{which.is.max}
\alias{fit.BN}
\alias{exhaustive_BN}
\alias{optimizemarginal}
\alias{ingreed_BN}
\alias{optimizecoregraph}
\alias{graychange}
\alias{is.dag}
\alias{optimizemarginal.nuInf}
\alias{parameters_continuous_Bayesian}
\alias{parameters_continuous_ML}
\alias{parameters_discrete_Bayesian}
\alias{parameters_discrete_ML}
\alias{score_continuous_Bayesian}
\alias{score_continuous_ML}
\alias{score_discrete_Bayesian}
\alias{score_discrete_ML}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{  internal functions  }
\description{
  internal functions: do not call these functions directly. 
}
\usage{
	various
}
%- maybe also 'usage' for other objects documented here.
\arguments{
	various
}
\details{
}
\value{
 	various
}
\author{ Holger Froehlich }
\keyword{models}

