\name{nem.bootstrap}
\alias{nem.bootstrap}

\title{Bootstrapping for nested effect models}
\description{Performs bootstrapping (resampling with replacement) on E-genes to assess the statistical stability of networks}
\usage{
nem.bootstrap(D,nboot=1000,inference="nem.greedy",models=NULL,type="mLL",para=NULL,hyperpara=NULL,Pe=NULL,Pm=NULL,Pmlocal=NULL,local.prior.size=length(unique(colnames(D))),local.prior.bias=1,triples.thrsh=0.5,lambda=0,delta=1,selEGenes=FALSE,verbose=TRUE)
}

\arguments{  
  \item{D}{data matrix with experiments in the columns (binary or continous)}
  \item{nboot}{number of bootstrap samples desired}
  \item{inference}{\code{search} to use exhaustive enumeration; or \code{triples} for triple-based inference; or \code{pairwise} for the pairwise heuristic; or \code{ModuleNetwork} for the module based inference; or \code{nem.greedy} for greedy hillclimbing}
  \item{models}{a list of adjacency matrices for model search}
  \item{type}{\code{mLL} or \code{FULLmLL} or \code{CONTmLL} or \code{CONTmLLDens} or \code{CONTmLLRatio}}
  \item{para}{vector of length two: false positive rate and false negative rate for non-binary data. Used by \code{mLL()}} 
  \item{hyperpara}{vector of length four: used by \code{FULLmLL()} for binary data}
  \item{Pe}{prior of effect reporter positions in the phenotypic hierarchy (same dimension as D)}
  \item{Pm}{prior over models (n x n matrix)}
  \item{Pmlocal}{local model prior for pairwise and triple learning. For pairwise learning generated by \code{local.model.prior()} according to arguments \code{local.prior.size} and \code{local.prior.bias}}
  \item{local.prior.size}{prior expected number of edges in the graph (for pairwise learning)}
  \item{local.prior.bias}{bias towards double-headed edges. Default: 1 (no bias; for pairwise learning)}    
  \item{triples.thrsh}{threshold for model averaging to combine triple models for each edge}  
  \item{lambda}{regularization parameter to include prior assumptions }
  \item{delta}{regularization parameter for automated E-gene subset selection (CONTmLLRatio only)}
  \item{selEGenes}{automated E-gene subset selection (includes tuning of delta for CONTmLLRatio)}
  \item{verbose}{do you want to see progression statements? Default: TRUE} 
}
\details{
  Calls \code{\link{nem}} or \code{\link{nemModelSelection}} internally, depending on whether or not lambda is a vector and Pm != NULL.
}
\value{
	matrix of bootstrap probabilities
}

\author{Holger Froehlich}


\seealso{code{\link{nem.calcSignificance}}, code{\link{nem}}}
\examples{
\dontrun{
   data("BoutrosRNAi2002")
   D <- BoutrosRNAiDiscrete[,9:16]
   p <- c(.13,.05)
   nem.bootstrap(D, para=p)            
}
}
\keyword{graphs}
\keyword{models}
