% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.rsphere.R
\name{normalize.rsphere}
\alias{normalize.rsphere}
\title{Normalize values on a sphere}
\usage{
normalize.rsphere(X, r = 1)
}
\arguments{
\item{r}{the diameter of a sphere, default 1.}

\item{x}{A vector or a matrix with values to be normalized.}
}
\value{
It returns a data frame with normalized values.
}
\description{
Normalizes values of a vector such that the sum of squared elements equal to $r^2$.
}
\author{
Marjan Cugmas
}
