% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq.sum.R
\name{seq.sum}
\alias{seq.sum}
\title{Generate sequnce of sums}
\usage{
\method{seq}{sum}(start, n.out, a)
}
\arguments{
\item{start}{The initial value.}

\item{n.out}{The number of generated values (length of a generated vector).}

\item{a}{what fraction of the previous value represents the current value.}
}
\value{
A vector.
}
\description{
Generates a vector of values. A $i$-th value is calculated as the $i-1$ value, multiplied by $a$.
}
\author{
Marjan Cugmas
}
