\name{set.default.parameters}
\alias{set.default.parameters}

\title{Get/set hyperparameters}
\description{Allows to set and retrieve various hyperparameters for different inference methods.}

\usage{
set.default.parameters(Sgenes, ...)
}
\arguments{
 \item{Sgenes}{character vector of S-gene identifiers}
 \item{...}{parameters to set (see details)}
}
\details{
Since version 2.5.4 functions in the nem package do not have any more a large amount of individual parameters. Instead there is just one hyperparameter, which is passed to all functions. Parameter values with the hyperparameter can be set with this function.
\describe{
 \item{type}{\code{mLL} or \code{FULLmLL} or \code{CONTmLL} or \code{CONTmLLBayes} or \code{CONTmLLMAP} or \code{depn}. \code{CONTmLLDens} and \code{CONTmLLRatio} are identical to \code{CONTmLLBayes} and \code{CONTmLLMAP} and are still supported for compatibility reasons. \code{mLL} and \code{FULLmLL} are used for binary data (see \code{BoutrosRNAiDiscrete}) and \code{CONTmLL} for a matrix of effect probabilities. \code{CONTmLLBayes} and \code{CONTmLLMAP} are used, if log-odds ratios, p-value densities or any other model specifies effect likelihoods. \code{CONTmLLBayes} refers to an inference scheme, were the linking positions of effect reporters to network nodes are integrated out, and \code{CONTmLLMAP} to an inference scheme, were a MAP estimate for the linking positions is calculated. \code{depn} indicates Deterministic Effects Propagation Networks (DEPNs).}
  \item{para}{vector of length two: false positive rate and false negative rate for binary data. Used by \code{mLL}} 
  \item{hyperpara}{vector of length four: used by \code{FULLmLL()} for binary data}
  \item{Pe}{prior of effect reporter positions in the phenotypic hierarchy (same dimension as D). Not used type \code{depn}. Default: NULL}
  \item{Pm}{prior over models (n x n matrix). Default: NULL}
  \item{Pm.frac_edges}{expected fraction of edges in the true S-gene graph}
  \item{Pmlocal}{local model prior for pairwise and triple learning. For pairwise learning generated by \code{local.model.prior} according to arguments \code{local.prior.size} and \code{local.prior.bias}}
  \item{local.prior.size}{prior expected number of edges in the graph (for pairwise learning). Default: no. nodes}
  \item{local.prior.bias}{bias towards double-headed edges. Default: 1 (no bias; for pairwise learning)}
  \item{triples.thrsh}{threshold for model averaging to combine triple models for each edge. Default: 0.5}
  \item{lambda}{regularization parameter to incorporate prior assumptions. May also be a vector of possible values, if \code{nemModelSelection} is used, Default: 0 (no regularization)}
  \item{delta}{regularization parameter for automated  subset selection of effect reporters. If no E-gene selection is wanted, set delta to 0. Default: 1/ (no. S-genes + 1)}
  \item{selEGenes.method}{If "regularization", E-gene selection is performed by introducing a "null" S-gene to which E-genes are attached with probability delta/ (no. S-genes + 1). If "iterative" and selEGenes=TRUE, \code{getRelevantEGenes} is called and a new model is trained on the selected E-genes. The process is then repeated until convergence. Default: "regularization"}
  \item{selEGenes}{Tune parameter delta for automated selection of E-genes. Default: FALSE. NOTE: Since version > 2.18.0 E-gene selection is now performed per default by using the regularization mechanism with parameter delta. If no E-gene selection is wanted, set delta to 0.}  
  \item{trans.close}{Should always transitive closed graphs be computed? Default: TRUE. NOTE: This has only an impact for type  \code{nem.greedyMAP} and \code{depn}. Default: TRUE}
  \item{backward.elimination}{For module networks and greedy hillclimbing inference: Try to eliminate edges increasing the likelihood. Works only, if trans.close=FALSE. Default: FALSE }
  \item{mode}{For Bayesian network inference and DEPNs: \code{binary_ML}: effects come from a binomial distribution - ML learning of parameters (Bayesian networks only); \code{binary_Bayesian}: effects come from a binomial distribution - Bayesian learning of parameters (Bayesian networks only); \code{continuous_ML}: effects come from a normal distribution - ML learning of parameters; \code{continuous_Bayesian}: effects come from a normal distribution - Bayesian learning of parameters.}
  \item{nu.intervention, lambda.intervention}{For \code{depn}: For any perturbed node we suppose the unknown mean mu given its unknown variance sigma2 to be drawn from N(nu.intervention, sigma2/lambda.intervention). Default: nu.intervention=0.6, lambda.intervention=4}
  \item{nu.no\_intervention, lambda.no\_intervention}{The same parameters for unperturbed nodes. Default: nu.no\_intervention=0.95, lambda.no\_intervention=4}
  \item{df.intervention, scale.intervention}{For \code{depn}: The unknown variance sigma2 for perturbed nodes is supposed to be drawn from Inv-\eqn{\chi^2}(df.intervention, scale.intervention). Default: df.intervention=4.4, scale.intervention=4.4}
  \item{df.no\_intervention, scale.no\_intervention}{The same parameters for unperturbed nodes. Default: df.no\_intervention=4.4, scale.no\_intervention=0.023}
  \item{map}{For \code{depn}: Mapping of interventions to network nodes. The format is a named list of strings with names being the interventions and entries being the network nodes. Default: Entries and names are the network nodes. }
  \item{outputdir}{Directory where to put diagnostic plots. Default: folder "QualityControl" in current working directory}
  \item{debug}{Print out or plot diagnostic information. Default: FALSE}
  \item{mc.cores}{number of cores to be used on a multicore processor. Default: 8}
  \item{mcmc.nsamples}{Number of MCMC samples to take. Default: 1e6}
  \item{mcmc.nburnin}{Number of additional samples for burnin phase. Default: 1e6} 
  \item{mcmc.seed}{random seed. Default: 1234}
  \item{mcmc.hyperprior}{Parameter for exponential distribution hyperprior for regularization parameter 1/lambda. Default: 1}
  \item{eminem.maxsteps}{Maximum number of iterations for the EM algorithm (MC.EMINEM). Default: 1000}
  \item{eminem.sdVal}{positive number, between 1 and ncol(D)*(ncol(D)-1): number of edges to change in one MCMC step; see paper for the author's choice. Default: 1}
  \item{eminem.changeHfreq}{positive number, mcmc.nsamples must be a multiple: the Empirical Bayes step is performed every <changeHfreq> steps (see paper for the author's choice); set >= mcmc.nsamples (or leave to default) to exclude this step. Default: NULL}
  \item{prob.cutoff}{Only edges with probability > prob.cutoff are assumed to be present. Default: 0.5}
 }
}
\value{
A list containing all parameters described above.
}
\examples{
control = set.default.parameters(LETTERS[1:5], type="CONTmLLBayes", selEGenes=TRUE) # set inference type and whether to use automatic E-gene selection for a network with nodes "A"-"E".
}
\references{

Markowetz, F.; Bloch, J. & Spang, R., Non-transcriptional Pathway Features Reconstructed from Secondary Effects of RNA interference. Bioinformatics, 2005, 21, 4026 - 4032\\

Markowetz, F.; Kostka, D.; Troyanskaya, O. & Spang, R., Nested Effects Models for High-dimensional Phenotyping Screens. Bioinformatics, 2007, 23, i305 - i312\\

Fr\"ohlich, H.; Fellmann, M.; S\"ultmann, H.; Poustka, A. & Beissbarth, T. Large Scale Statistical Inference of Signaling Pathways from RNAi and Microarray Data. BMC Bioinformatics, 2007, 8, 386\\

Fr\"ohlich, H.; Fellmann, M.; S\"ultmann, H.; Poustka, A. & Beissbarth, T. Estimating Large Scale Signaling Networks through Nested Effect Models with Intervention Effects from Microarray Data. Bioinformatics, 2008, 24, 2650-2656\\ 

Tresch, A. & Markowetz, F., Structure Learning in Nested Effects Models Statistical Applications in Genetics and Molecular Biology, 2008, 7\\

Zeller, C.; Fr\"ohlich, H. & Tresch, A., A Bayesian Network View on Nested Effects Models EURASIP Journal on Bioinformatics and Systems Biology, 2009, 195272\\

Fr\"ohlich, H.; Tresch, A. & Beissbarth, T., Nested Effects Models for Learning Signaling Networks from Perturbation Data. Biometrical Journal, 2009, 2, 304 - 323\\

Fr\"ohlich, H.; Sahin, \"O.; Arlt, D.; Bender, C. & Beissbarth, T. Deterministic Effects Propagation Networks for Reconstructing Protein Signaling Networks from Multiple Interventions. BMC Bioinformatics, 2009, 10, 322\\

Fr\"ohlich, H.; Praveen, P. & Tresch, A., Fast and Efficient Dynamic Nested Effects Models. Bioinformatics, 2011, 27, 238-244\\

Niederberger, T.; Etzold, S.; Lidschreiber, M; Maier, K.; Martin, D.; Fr\"ohlich, H.; Cramer, P.; Tresch, A., MC Eminem Maps the Interaction Landscape of the Mediator, PLoS Comp. Biol., 2012, submitted. 

}
\author{Holger Froehlich}
\keyword{graphs}
\keyword{models}
