\name{subsets}
\alias{subsets}

\title{Subsets}
\description{
  subsets
}
\usage{
subsets(n, r, v = 1:n, set = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n}{bli}
  \item{r}{bla}
  \item{v}{blo}
  \item{set}{blu}
}

\details{
taken from the programmers corner of some R-News issue by Dennis
}

\value{
\item{n}{bli}
  \item{r}{bla}
  \item{v}{blo}
 
}

\author{Dennis Kostka <URL: http://www.molgen.mpg.de/~kostka>}



%\seealso{}
\examples{
 ## bla
}
\keyword{graphs}
