% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARI.R
\name{MARI}
\alias{MARI}
\title{Modified Rand Index (MRI) and Modified Adjusted Rand Index (MARI)}
\usage{
MARI(U = U, V = V, k = 1000)
}
\arguments{
\item{U}{The first partition.}

\item{V}{The second partition.}

\item{k}{the number of iterations to estimate the expected value in the case of two random and independent partitions.}
}
\value{
A single value.
}
\description{
It computes the value of the adjuster or non-adjusted Modified Rand Index.
}
\details{
The corresponding values for the first and/or second partitions for in-comers and/or out-goings have to be NA.
}
\references{
Cugmas, M., & Ferligoj, A. (2018). Comparing two partitions of non-equal sets of units. Advances in Methodology and Statistics 1(15), 1-23.
}
\author{
Marjan Cugmas
}
