\name{nem}
\alias{nem}
\alias{plot.nem}
\alias{print.nem}

\title{Nested Effects Models - main function}
\description{The main function to infer a phenotypic hierarchy from data}
\usage{
nem(D,inference="pairwise",models=NULL,type="mLL",para=NULL,hyperpara=NULL,Pe=NULL,Pm=NULL,local.prior.size=length(unique(colnames(D))),local.prior.bias=1,verbose=TRUE)
}

\arguments{
  \item{D}{binary data matrix with experiments in the columns}
  \item{inference}{\code{search} by function \code{score()}; or \code{pairwise} to use function \code{pairwise.posterior()}}
  \item{models}{a list of adjacency matrices for model search. If NULL, \code{enumerate.models} is used for exhaustive enumeration of all possible models.}
  \item{type}{\code{mLL} or \code{FULLmLL}}
  \item{para}{vector of length two: false positive rate and false negative rate. Used by \code{mLL()}} 
  \item{hyperpara}{vector of length four: used by \code{FULLmLL()}}
  \item{Pe}{prior of effect reporter positions in the phenotypic hierarchy}
  \item{Pm}{prior over models. For pairwise learning generated by \code{local.model.prior()} according to arguments \code{local.prior.size} and \code{local.prior.bias}}
  \item{local.prior.size}{prior expected number of edges in the graph}
  \item{local.prior.bias}{bias towards double-headed edges. Default: 1 (no bias)}
  \item{verbose}{do you want to see progression statements" Default: TRUE}
}
\details{
  \code{nem} is an interface to the functions \code{score()} and \code{pairwise.posterior()}. 
  
  \code{plot.nem} plots the inferred phenotypic hierarchy as a directed graph, and \code{print.nem} gives an overview over the 'nem' object.
  
  
}
\value{
An object of class 'score' or class 'pairwise' containing slots
\item{graph}{the inferred phenotypic hierarchy}
\item{pos}{posterior distribution of positions of effect reporters}
\item{mappos}{estimated position of effects in the phenotypic hierarchy}
and additional ones according to the function used for inference.
}

\author{Florian Markowetz <URL: http://genomics.princeton.edu/~florian>}


\seealso{\code{\link{score}}, \code{\link{pairwise.posterior}}, \code{\link{local.model.prior}}, \code{\link{enumerate.models}}}
\examples{
   data("BoutrosRNAi2002")
   D <- BoutrosRNAiDiscrete[,9:16]
   res1 <- nem(D,para=c(.13,.05),inference="search")
   res2 <- nem(D,para=c(.13,.05),inference="pairwise")

   par(mfrow=c(1,2))
   plot(res1,main="by exhaustive search")
   plot(res2,main="by pairwise heuristic")
   
}
\keyword{graphs}
\keyword{models}
